UHHthesis: R Markdown thesis templates designed for the University of Hamburg (UHH)

A package for creating Bachelor and Masters theses in PDF and MS Word format using bookdown and R Markdown.

thesis_pdf_en

Creates a PDF/LaTeX thesis document in English

thesis_word_en

Creates a Microsoft Word (docx) thesis document in English

thesis_pdf_de

Creates a PDF/LaTeX thesis document in German

thesis_word_de

Creates a Microsoft Word (docx) thesis document in German

See also

Author

Maintainer: Saskia Otto saskia.a.otto@gmail.com (ORCID)